requirements.txt
Flask==2.0.1
pypubsub==4.0.3
sounddevice==0.4.6
soundfile==0.12.1
simpleaudio==1.0.4
openai==0.27.10
colorama==0.4.6
numpy==1.25.2
pydub==0.25.1
python-dotenv==1.0.0
pyaudio==0.2.13
